% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{gen_chisq}
\alias{gen_chisq}
\title{Chi-Squared Points}
\usage{
gen_chisq(n, df, m = 10)
}
\arguments{
\item{n}{number of points to generate.}

\item{df}{degrees-of-freedom (must be greater than 2).}

\item{m}{positive scalar such that n points are selected from an
initial set of (m * n) points by Stein Thinning.}
}
\value{
n-vector containing the sampled selected points.
}
\description{
Generates points targeting a chi-squared distribution.
}
