% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{gen_exp}
\alias{gen_exp}
\title{Exponential Points}
\usage{
gen_exp(n, rate = 1)
}
\arguments{
\item{n}{number of points to generate.}

\item{rate}{rate parameter.}
}
\value{
n-vector containing the sampled points.
}
\description{
Generates points targeting an Exponential distribution.
}
