% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{gen_gh}
\alias{gen_gh}
\title{Tukey's g-and-h Points}
\usage{
gen_gh(n, a = 0, b = 1, g = 0, h = 0)
}
\arguments{
\item{n}{number of points to generate.}

\item{a}{location parameter.}

\item{b}{scale parameter (> 0).}

\item{g}{skewness parameter.}

\item{h}{tail-shape parameter (>= 0).}
}
\value{
n-vector containing the sampled points.
}
\description{
Generates points targeting a g-and-h distribution.
}
