% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{gen_gk}
\alias{gen_gk}
\title{g-and-k Points}
\usage{
gen_gk(n, a = 0, b = 1, g = 0, k = 0)
}
\arguments{
\item{n}{number of points to generate.}

\item{a}{location parameter.}

\item{b}{scale parameter (> 0).}

\item{g}{skewness parameter.}

\item{k}{tail-shape parameter (k > -0.5).}
}
\value{
n-vector containing the sampled points.
}
\description{
Generates points targeting a g-and-k distribution.
}
