% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{gen_norm}
\alias{gen_norm}
\title{Normal Points}
\usage{
gen_norm(n, mu = 0, sig = 1, m = 10)
}
\arguments{
\item{n}{number of points to generate.}

\item{mu}{mean.}

\item{sig}{standard deviation.}

\item{m}{positive scalar such that n points are selected from an
initial set of (m * n) points by Stein Thinning.}
}
\value{
n-vector containing the sampled selected points.
}
\description{
Generates points targeting a normal distribution.
}
