% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{gen_unif}
\alias{gen_unif}
\title{Uniform Points}
\usage{
gen_unif(n, l = 0, u = 1)
}
\arguments{
\item{n}{number of points to generate.}

\item{l}{lower bound.}

\item{u}{upper bound.}
}
\value{
n-vector containing the sampled points.
}
\description{
Generates points targeting a uniform distribution.
}
