% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{grad_chisq}
\alias{grad_chisq}
\title{Chi-Squared Gradient}
\usage{
grad_chisq(x, df)
}
\arguments{
\item{x}{point at which the gradient is computed.}

\item{df}{degrees-of-freedom.}
}
\value{
gradient of the log of the chi-squared PDF.
}
\description{
The logarithmic derivative of the chi-squared density function.
}
