% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient.R
\name{grad_norm}
\alias{grad_norm}
\title{Normal Gradient}
\usage{
grad_norm(x, mu = 0, sig = 1)
}
\arguments{
\item{x}{point at which the gradient is computed.}

\item{mu}{mean.}

\item{sig}{standard deviation.}
}
\value{
gradient of the log of the normal PDF.
}
\description{
The logarithmic derivative of the normal density function.
}
