% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{Test}
\usage{
test()
}
\value{
list of matrices of sample moments:
\itemize{
\item ss - first four moments of samples from \code{gen_norm}.
\item id - first four moments of samples from \code{rnorm}.
}
}
\description{
Plots the moments of the generated samples.
}
